Fix bug in VCFHeader that caused newly-added header lines to be silently dropped
There was a nasty bug in VCFHeader that caused newly-added header lines
to only be added to type-specific lookup tables and not to the master list
of header lines. This meant that new lines got silently dropped when writing
a modified VCFHeader back out via VCFWriter.

Fixed the issue and added comprehensive unit tests for VCFHeader.addMetaDataLine(),
as well as a VCFWriter test to ensure new header lines get persisted.

Also attempted to improve documentation and method naming a bit in an
effort to avoid future bugs like this. Eg., there was a method "loadVCFVersion()"
that actually REMOVED the vcf version lines!