SimpleBEDFeature moved into FullBEDFeature because of uselessness and because he's just breaking meaning of class heritage
The SimpleBEDFeature class is an unnecessary layer between interface and the actual class. There's no reason of having that non-functional class. There's no reason of having the method which returns a list of objects-of-nested-subclass-of-superclass. That seems to be an architecture mistake. And also it forces to implement method getExons which actually must not be in "Simple" just by sense of that class.
Additionally (to get rid of usage of successor`s-subclass into superclass) I've created FeatureSubSequence interface and added "implements" for both Exon and Exon2. No logic was changed here -- only a heritage model.
Also overkilling and heavy-regexp-compiling String#replaceAll(String, String) replaced with String#repalce(CharSequence, CharSequence)