Fix self-suppression problem with exception handling in AbstractAsyncWriter
An exception thrown during synchronousWrite() by the underlying thread
is typically picked up during write()'s invocation of checkAndRethrow().
Subsequent calling of close() (e.g. by try-with-resources) would result
in an attempt to throw the same exception instance.

This commit ensures that the underlying exception is only thrown once,
by clearing the AtomicReference to the Throwable before throwing.

Includes a unit test which demonstrates the problem.