adding CloseableIterator.asyncMap(f)
this returns a new iterator which applies f to the contents of the
previous iterator asyncronously using a threadpool

It pulls the underlying iterator repeatedly to build chunks, and then
executes those chunks asyncronously on a thread pool.  The
implementation is pretty rough and could probably be made more
efficient. (and no tests are attached... use at your
own risk...)