Implement Read interface
Create common Read interface to move to a interface-based API in htsjdk.
The new Read interface is packed in htsjdk.samtools.read to add more
structure to the API. The interface is based in the current SAMRecord
and in the GATK's read interface, keeping some ideas from both.

In addition to the new interface, some more classes were added:

* ReadConstants: include constants for the Read interface. Correspond to
  some constants in the SAMRecord
* ReadAttribute: another API interface to allow different
  representations of the read attributes.
* ReadUtils: utility class for working with the read interface in a
  implementation-independent way. Include few examples of no-API methods
  for getting un-clipped start/end.