Support BAMs with >65535 CIGAR operations
This is the solution that @yfarjoun and I have agreed on. In the presence of
CIGAR with >65535 operators, we move the real CIGAR to the CG tag and put a
fake full clipping CIGAR (i.e. <readLength>S) in place.

`./gradlew` reported one test failure due to `EnaRefServiceTest`. Looking at
the call stack, I don't think that is my fault. The rest of tests have passed.

On an example SAM and BAM (from <http://lh3lh3.users.sourceforge.net/data>),
`PrintReadsExample` gives the desired BAM output when taking both SAM and BAM
as input. Both shallow and deep decoding also write correct BAMs.