Prevent NPE in BlockCompressedInputStream exception messages(#1067)
Several exception
messages in BlockCompressedInputStream use mFile.getSource() in the message, but mFile can be null when the input is a stream.  This caused NPE's which interrupted the exceptions.

Replaced those calls with a new safe method which does a null check and returns "data_stream" when mFile is null.