Support for block-compressed indexed FASTA references (#1014)
* HTSJDK can now read and query block compressed fasta files that have a .fai and .gzi index with them.  

* Extract AbstractIndexedFastaSequenceFile

Extract a common class for indexed FASTA sequence file, to use a
different random-access like BlockCompressedInputStream instead of
SeekableByteChannel.

In addition, the IndexedFastaSequenceFile throws for block-compressed
references.

* Add BlockCompressedIndexedFastaSequenceFile

New class for indexed FASTA reference sequences to support
block-compressed input. In addition, the factory method was updated to
return the new class in the case of block-compressed files.