VCFWriter accepts Path
and associated changes necessary for this to happen.

I also added a Path test to VariantContextWriterBuilderUnitTest
(using JimFS).

The benefit of this change is that a Path can point
to files on the cloud (e.g. via GCloud's NIO
Filesystem Provider for Google Cloud Storage).

With this change, I was able to change GATK to write VCF output
to the cloud:

$ ./gatk GatherVcfsCloud -I ./sample.vcf -O gs://jps_bucket/out.vcf