Changed IntervalList fromFiles() so that it doesn't call .unique() on the interval list before returning it. While this is a change in API, it is more consistent with fromFile() that similarly doesn't call .unique() before returning the result.
This should fix an issue in picard where CollectHsMetrics (incorrectly) merges abutting intervals prior to merging them.

Users who would like to keep the old behaviour should call uniqued() on the result prior to using.

Put in the foundation of a more flexible way to uniquify intervals, with a combiner class whose properties can be set.