Adding a code of conduct document for htsjdk.
This is the same code of conduct that was agreed on for htsjdk-next so hopefully it won't be to controversial here.

I think it's good to state that we expect participants to be respectful of each other
and that part of the responsibility of the maintainers is to maintain a positive and welcoming
community as well as just functional code.