Zero len array (#1674)
* Remove setUnsignedArrayAttribute zero-length-array check

This function was omitted from PR #1194. Zero-length 'B' arrays
are valid, so remove this check too and add a test case.

* Remove Slice setAttribute zero-length-array check

Slices' optional tags follow the rules for BAM tagged fields, so
zero-length 'B' arrays are valid. Slice tags are currently unused,
so the question is somewhat moot and there are no test cases.

* removed unused setAttribute methods

---------

Co-authored-by: John Marshall <jmarshall@hey.com>